[Refactor] Shifted Projects saving data to ProjectRepository
