+++ src/main/java/repositories/ProjectRepository.java
@@ -1,18 +1,26 @@
 package repositories;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
+import controllers.ConsoleInputController;
 import models.data.IProject;
 import models.data.Project;
 import util.factories.ProjectFactory;
 import util.log.DukeLogger;
 
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.util.ArrayList;
 
 public class ProjectRepository implements IRepository<Project> {
     private ArrayList<Project> allProjects;
     private ProjectFactory projectFactory;
+    private String filePath = System.getProperty("user.dir") + "/savedProjects.json";
 
     public ProjectRepository() {
-        this.allProjects = new ArrayList<>();
+        loadProjectsData();
         this.projectFactory = new ProjectFactory();
     }
 
@@ -31,6 +39,7 @@ public class ProjectRepository implements IRepository<Project> {
         }
         Project newlyCreatedProject = (Project) newProject;
         allProjects.add(newlyCreatedProject);
+        saveProjectsData();
         return true;
     }
 
@@ -51,8 +60,10 @@ public class ProjectRepository implements IRepository<Project> {
     public boolean deleteItem(int projectNumber) {
         try {
             this.allProjects.remove(projectNumber - 1);
+            saveProjectsData();
             return true;
         } catch (IndexOutOfBoundsException err) {
+            saveProjectsData();
             return false;
         }
     }
@@ -99,5 +110,36 @@ public class ProjectRepository implements IRepository<Project> {
             toPrintAll.add(toPrint);
         }
         return toPrintAll;
+    }
+
+    private void saveProjectsData() {
+        Gson gson = new GsonBuilder()
+                .setPrettyPrinting()
+                .create();
+        try {
+            DukeLogger.logDebug(ConsoleInputController.class, "Saving to file.");
+            FileWriter fileWriter = new FileWriter(filePath);
+            gson.toJson(this.allProjects, fileWriter);
+            fileWriter.flush();
+            fileWriter.close();
+            DukeLogger.logDebug(ConsoleInputController.class, "File saved.");
+        } catch (IOException err) {
+            DukeLogger.logError(ConsoleInputController.class, "savedProjects file is not found.");
+        }
+    }
+
+    private void loadProjectsData() {
+        Gson gson = new Gson();
+        try (FileReader fileReader = new FileReader(filePath)) {
+            DukeLogger.logDebug(ConsoleInputController.class, "Loading saved file.");
+            this.allProjects = gson.fromJson(fileReader, new TypeToken<ArrayList<Project>>(){}.getType());
+            if (this.allProjects == null) {
+                this.allProjects = new ArrayList<>();
+            }
+            DukeLogger.logDebug(ConsoleInputController.class, "Saved file loaded.");
+        } catch (IOException err) {
+            DukeLogger.logError(ConsoleInputController.class, "Saved file not loaded");
+            this.allProjects = new ArrayList<>();
+        }
     }
 }
